+++ vars.sh
@@ -33,7 +33,7 @@ PATCH=`./vendor/semver get patch ${VERSION}`
 
 # Versions for container images and helm charts
 PRODUCT_CATALOG_UPDATE_VERSION=1.3.2
-UAN_CONFIG_VERSION=1.12.8
+UAN_CONFIG_VERSION=1.14.1
 
 # Versions for UAN images
 UAN_IMAGE_RELEASE=stable
@@ -44,10 +44,11 @@ UAN_IMAGE_NAME=cray-application-sles15sp4.$UAN_IMAGE_ARCH-$UAN_IMAGE_VERSION
 UAN_IMAGE_URL=https://artifactory.algol60.net/artifactory/csm-images/$UAN_IMAGE_RELEASE/application
 
 # Dependencies for UAIs on Application nodes
-K3S_VERSION=1.26.0
-METALLB_VERSION=0.13.7
-HAPROXY_VERSION=1.17.3
-HAPROXY_CONTAINER_VERSION=2.6.6
+K3S_VERSION=1.27.3
+METALLB_VERSION=0.13.10
+HAPROXY_VERSION=1.19.2
+FRR_VERSION=8.4.2
+HAPROXY_CONTAINER_VERSION=2.8.1
 K3S_URL=https://github.com/k3s-io/k3s/releases/download/v$K3S_VERSION%2Bk3s1
 K3S_INSTALLER=https://get.k3s.io
 METALLB_URL=https://metallb.github.io/metallb